requirements.txt
python-telegram-bot==20.3
scikit-learn
nltk
pandas
python-dotenv
requests